+++ mods/BrewLAN_Plenae/CityGenerationAeon/mod_info.lua
@@ -1,8 +1,8 @@
-name = "Cityscapes"
-uid = "190261d0-4bb0-11e2-bcSC-CITIES000001"
-version = 1
+name = "Cityscapes: Aeon"
+uid = "190261d0-4bb0-11e2-bcSC-CITIESA00001"
+version = 0
 copyright = "© 2020–2020 Sean Wheeldon"
-description = "Randomly generates UEF civilian cities."
+description = "Requires Cityscapes. Adds the potential for cities with Aeon structures."
 author = "Balthazar"
 icon = "/mods/BrewLAN_Plenae/CityGeneration/mod_icon.png"
 selectable = true